Cambio tipos de datos de importes float a BigDecimal

+++ TurboERP_backend/src/main/java/com/turbomaquinas/DAO/comercial/PagosRM.java
@@ -16,7 +16,7 @@ public class PagosRM implements RowMapper<Pagos> {
 		p.setId(rs.getInt("id"));
 		p.setFolio(rs.getInt("folio"));
 		p.setFecha_pago(rs.getDate("fecha_pago"));
-		p.setImporte(rs.getFloat("importe"));
+		p.setImporte(rs.getBigDecimal("importe"));
 		p.setTipo_cambio_cliente(rs.getFloat("tipo_cambio_cliente"));
 		p.setFecha_baja(rs.getDate("fecha_baja"));
 		p.setMes_baja(rs.getInt("mes_baja"));
